@@ -4,10 +4,10 @@
         "friendsofphp/php-cs-fixer": "^3.89.2"
     },
     "config": {
-        "bump-after-update": true,
         "platform": {
-            "php": "8.2"
+            "php": "8.4"
         },
+        "bump-after-update": true,
         "sort-packages": true
     }
 }
