@@ -3,6 +3,7 @@
     "require": {
         "phpstan/extension-installer": "^1.4.3",
         "phpstan/phpstan": "^2.1.32",
+        "phpstan/phpstan-deprecation-rules": "^2.0.3",
         "phpstan/phpstan-symfony": "^2.0.8"
     },
     "config": {
@@ -11,7 +12,7 @@
         },
         "bump-after-update": true,
         "platform": {
-            "php": "8.2"
+            "php": "8.4"
         },
         "sort-packages": true
     }
