@@ -1,13 +1,13 @@
 {
     "type": "project",
     "require": {
-        "friendsofphp/php-cs-fixer": "^3.89.2"
+        "vincentlanglet/twig-cs-fixer": "^3.10.0"
     },
     "config": {
-        "bump-after-update": true,
         "platform": {
-            "php": "8.2"
+            "php": "8.4"
         },
+        "bump-after-update": true,
         "sort-packages": true
     }
 }
